requirements.txt
keras==3.10.0
numpy==2.1.3
opencv_python==4.11.0.86
Pillow==11.2.1
matplotlib==3.10.3
pandas==2.3.0
pydot==4.0.0
scikit-learn==1.7.0
graphviz==0.20.3